requirements_dev.txt
-r requirements.txt

django-debug-toolbar==1.4
ipdb==0.10.2